Added IR visualization JSON export code.
There is a new delitec option `--list-transformers`
that prints out all transformers scheduled for compilation and
exits without compiling. I added `toString` methods for the Delite
tranformers so these print out nicely in this list.

The JSON export is hardcoded right now in `main` in
`DeliteApplication.scala` as a transformer that is scheduled after
all other tranformers. It is also possible to schedule the JSON
export at other times using `DeliteIRGraphExport.scala`.