Adding the "HwGen" file that contains the skeleton of the hardware backend. The change is described in the comments and below: 1. Delite's architecture for a backend is a collection of traits, where each trait overrides the 'emitNode' method and defines the translation for a few kinds of IR nodes. For every other non-matched IR node, the next trait's emitNode (in linearization order) is called using super.emitNode().
2. So far, the codegen logic is sprinkled over several files, both in LMS and Delite.
In case of the hardware backend here, an effort has been made to unify all
codegen logic into a single package and file so that it is easier to read.
Except for a minor change in LMS (adding a new Value in the Targets object in
AccModel.scala), the entire hardware backend is contained within Delite.
This means that code to generate hardware for basic LMS nodes is also within
Delite.

3. A basic subset of traits have been identified (inspired from the C backend)
and skeletons declared. The backend is still a collection of traits, just
like the others.

4. In order to use the backend, the DSL/test program has to:
    - mix-in all the traits together into a supertrait
    - Implement the getCodeGenPkg function to instantiate the code gen