Bring back immutable clone opt for DeliteArray as a different function.
Optimizing away clones for immutable arrays was first added in
c301618a2086b862a15bb099812e60195fed7240 and then removed in
a216ecc254b3555376fd61253e469be2456656c4 because it was not always
safe. Here we bring it back (with improved checking including mutable
aliases), but under a separate method so that it is not the default.